Add DateTime element to avoid code duplication
