+++ src/modules/blocks/style.pcss
@@ -20,6 +20,10 @@
 		text-decoration: none;
 		color: #666;
 	}
+
+	.tribe-element__datetime {
+		display: inline-block;
+	}
 }
 
 .tribe-editor-events-details {
