+++ src/modules/elements/index.js
@@ -1,2 +1,3 @@
 export { default as TermsList } from './terms-list'
 export { default as OrganizerForm } from './organizer-form'
+export { default as DateTime } from './datetime'
